order-service/requirements.txt
# Common dependencies
grpcio==1.62.0
grpcio-tools==1.62.0
protobuf==4.25.0
sqlmodel==0.0.16
sqlalchemy==2.0.23

# Database
aiosqlite==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
